requirements.txt
haystack-ai
qdrant-haystack
pypdf
python-docx
azure-identity
azure-storage-file-share
python-dotenv
requests
